requirements.txt
pandas as pd
plotly.express as px
streamlit as st
pickle as pkl
numpy as np
matplotlib.pyplot as plt